move all unnessesary postfix
