+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/assets/catalog.py
