+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/assets/spec_cache.py
